Change to separate the definition of a phase and its scheduling. Now phases are defined and *then* scheduled to run in either the Collector, Mutator, or Global context. This is much more natural and follows from the transition from Locals to Mutator and Collector contexts. This should make the ordering and execution of phases much more natural and obvious.

+++ MMTk/src/org/mmtk/plan/CollectorContext.java
@@ -154,7 +154,7 @@ import org.vmmagic.unboxed.*;
    * @param primary Should this thread be used to execute any single-threaded
    * local operations?
    */
-  public abstract void collectionPhase(int phaseId, boolean primary);
+  public abstract void collectionPhase(short phaseId, boolean primary);
 
   /** @return The current trace instance. */
   public abstract TraceLocal getCurrentTrace();
