+++ MMTk/src/org/mmtk/plan/MutatorContext.java
@@ -109,7 +109,7 @@ import org.vmmagic.unboxed.*;
    * @param primary Should this thread be used to execute any single-threaded
    * local operations?
    */
-  public abstract void collectionPhase(int phaseId, boolean primary);
+  public abstract void collectionPhase(short phaseId, boolean primary);
 
   /****************************************************************************
    *
