+++ MMTk/src/org/mmtk/plan/copyms/CopyMS.java
@@ -92,7 +92,7 @@ import org.vmmagic.pragma.*;
    * @param phaseId Collection phase to execute.
    */
   @Inline
-  public final void collectionPhase(int phaseId) {
+  public final void collectionPhase(short phaseId) {
     if (phaseId == PREPARE) {
       super.collectionPhase(phaseId);
       trace.prepare();
