+++ MMTk/src/org/mmtk/plan/copyms/CopyMSCollector.java
@@ -111,7 +111,7 @@ import org.vmmagic.unboxed.*;
    * @param primary Use this thread for single-threaded local activities.
    */
   @Inline
-  public final void collectionPhase(int phaseId, boolean primary) {
+  public final void collectionPhase(short phaseId, boolean primary) {
     if (phaseId == CopyMS.PREPARE) {
       super.collectionPhase(phaseId, primary);
       mature.prepare();
