+++ MMTk/src/org/mmtk/plan/generational/copying/GenCopy.java
@@ -133,7 +133,7 @@ import org.vmmagic.pragma.*;
    * @param phaseId Collection phase to process
    */
   @Inline
-  public void collectionPhase(int phaseId) {
+  public void collectionPhase(short phaseId) {
     if (traceFullHeap()) {
       if (phaseId == PREPARE) {
         super.collectionPhase(phaseId);
