+++ MMTk/src/org/mmtk/plan/generational/copying/GenCopyCollector.java
@@ -124,7 +124,7 @@ import org.vmmagic.pragma.*;
    * @param primary True if this thread should peform local single-threaded
    * actions.
    */
-  public void collectionPhase(int phaseId, boolean primary) {
+  public void collectionPhase(short phaseId, boolean primary) {
     if (global().traceFullHeap()) {
       if (phaseId == GenCopy.PREPARE) {
         super.collectionPhase(phaseId, primary);
