+++ MMTk/src/org/mmtk/plan/semispace/gctrace/GCTraceCollector.java
@@ -65,7 +65,7 @@ import org.vmmagic.pragma.*;
    * @param phaseId The collection phase to perform
    * @param primary perform any single-threaded local activities.
    */
-  public void collectionPhase(int phaseId, boolean primary) {
+  public void collectionPhase(short phaseId, boolean primary) {
     if (phaseId == GCTrace.START_CLOSURE) {
       inducedTrace.startTrace();
       return;
